Moved StoredAddress and StoredValue from Dependency.[cpp,h] to TxValues.[cpp,h]
