+++ lib/Core/TxTree.cpp
@@ -1065,7 +1065,7 @@ bool SubsumptionTableEntry::subsumed(
               msg1 += " (with successful memory bound checks)";
             }
 
-            it2->first->loc->print(stream, padding);
+            it2->first->print(stream, padding);
             stream.flush();
             klee_message("#%lu=>#%lu: Check failure as memory region in the "
                          "table does not exist in the state%s:\n%s",
@@ -1153,7 +1153,7 @@ bool SubsumptionTableEntry::subsumed(
                   std::string msg3;
                   llvm::raw_string_ostream stream1(msg3);
 
-                  it2->first->loc->print(stream1, makeTabs(1));
+                  it2->first->print(stream1, makeTabs(1));
                   stream1.flush();
 
                   klee_message("with value stored in address:\n%s",
@@ -1180,7 +1180,7 @@ bool SubsumptionTableEntry::subsumed(
                 std::string msg3;
                 llvm::raw_string_ostream stream1(msg3);
 
-                it2->first->loc->print(stream1, makeTabs(1));
+                it2->first->print(stream1, makeTabs(1));
                 stream1.flush();
 
                 klee_message("with value stored in address:\n%s", msg3.c_str());
@@ -1190,7 +1190,7 @@ bool SubsumptionTableEntry::subsumed(
         }
 
         if (!stateSymbolicMap.empty()) {
-          const ref<Expr> tabledConcreteOffset = it2->first->loc->getOffset();
+          const ref<Expr> tabledConcreteOffset = it2->first->getOffset();
           ref<Expr> conjunction;
 
           for (Dependency::SymbolicStoreMap::const_iterator
@@ -1200,10 +1200,10 @@ bool SubsumptionTableEntry::subsumed(
 
             // We make sure the context part of the addresses (the allocation
             // site and the call history) are equivalent.
-            if (it2->first->loc->getContext() != it3->first->loc->getContext())
+            if (it2->first->getContext() != it3->first->getContext())
               continue;
 
-            ref<Expr> stateSymbolicOffset = it3->first->loc->getOffset();
+            ref<Expr> stateSymbolicOffset = it3->first->getOffset();
             ref<Expr> newTerm;
 
             stateValue = it3->second;
@@ -1306,7 +1306,7 @@ bool SubsumptionTableEntry::subsumed(
                it2 = tabledSymbolicMap.begin(),
                ie2 = tabledSymbolicMap.end();
            it2 != ie2; ++it2) {
-        ref<Expr> tabledSymbolicOffset = it2->first->loc->getOffset();
+        ref<Expr> tabledSymbolicOffset = it2->first->getOffset();
         ref<StoredValue> tabledValue = it2->second;
 
         for (Dependency::ConcreteStoreMap::const_iterator
@@ -1316,10 +1316,10 @@ bool SubsumptionTableEntry::subsumed(
 
           // We make sure the context part of the addresses (the allocation
           // site and the call history) are equivalent.
-          if (it2->first->loc->getContext() != it3->first->loc->getContext())
+          if (it2->first->getContext() != it3->first->getContext())
             continue;
 
-          ref<Expr> stateConcreteOffset = it3->first->loc->getOffset();
+          ref<Expr> stateConcreteOffset = it3->first->getOffset();
           ref<StoredValue> stateValue = it3->second;
           ref<Expr> newTerm;
 
@@ -1378,10 +1378,10 @@ bool SubsumptionTableEntry::subsumed(
 
           // We make sure the context part of the addresses (the allocation
           // site and the call history) are equivalent.
-          if (it2->first->loc->getContext() != it3->first->loc->getContext())
+          if (it2->first->getContext() != it3->first->getContext())
             continue;
 
-          ref<Expr> stateSymbolicOffset = it3->first->loc->getOffset();
+          ref<Expr> stateSymbolicOffset = it3->first->getOffset();
           ref<StoredValue> stateValue = it3->second;
           ref<Expr> newTerm;
 
@@ -1786,7 +1786,7 @@ void SubsumptionTableEntry::print(llvm::raw_ostream &stream,
         if (it1 != is1 || it2 != is2)
           stream << tabsNext << "------------------------------------------\n";
         stream << tabsNext << "address:\n";
-        it2->first->loc->print(stream, tabsNextNext);
+        it2->first->print(stream, tabsNextNext);
         stream << "\n";
         stream << tabsNext << "content:\n";
         it2->second->print(stream, tabsNextNext);
@@ -1809,7 +1809,7 @@ void SubsumptionTableEntry::print(llvm::raw_ostream &stream,
         if (it1 != is1 || it2 != is2)
           stream << tabsNext << "------------------------------------------\n";
         stream << tabsNext << "address:\n";
-        it2->first->loc->print(stream, tabsNextNext);
+        it2->first->print(stream, tabsNextNext);
         stream << "\n";
         stream << tabsNext << "content:\n";
         it2->second->print(stream, tabsNextNext);
